test/forks: rename tests
Just like the regular bitcoinjs-lib tests have `bitcoin` and `testnet`
networks in the same test files, we should group `$forkCoin` and
`$forkCoinTest` under the mainnet name.

Rename `btg` to the network name `bitcoingold`.
