+++ test/forks/bitcoingold/transaction_builder.js
